refactor(pacts): extract crowd contracts from legacy
Move the Crowd (user-identity) cluster out of the catch-all
pacts/legacy.py into a dedicated pacts/crowd.py: UserType, UserDTO,
UserData, UserRepositoryProtocol, ConnectedUserRepositoryProtocol.
These form a closed cluster depending only on pydantic/typing/enum.

Update every import site to source these names directly from
ludamus.pacts.crowd (no crowd re-exports added to pacts/__init__.py,
per project convention). Pure move, no behavior change.
