+++ src/ludamus/mills/legacy.py
@@ -44,14 +44,16 @@ from ludamus.pacts import (
     TrackDTO,
     UnitOfWorkProtocol,
     UploadedFileProtocol,
-    UserData,
-    UserDTO,
     UserEnrollmentConfigData,
     UserEnrollmentConfigDTO,
-    UserRepositoryProtocol,
-    UserType,
     VirtualEnrollmentConfig,
     WizardData,
+)
+from ludamus.pacts.crowd import (
+    UserData,
+    UserDTO,
+    UserRepositoryProtocol,
+    UserType,
 )
 from ludamus.specs.encounter import ENCOUNTER_DEFAULT_DURATION
 from ludamus.specs.proposal import PROPOSAL_RATE_LIMIT_SECONDS
