+++ tests/integration/web/chronology/test_event_anonymous_activate_action.py
@@ -4,7 +4,7 @@ from django.contrib import messages
 from django.urls import reverse
 
 from ludamus.adapters.db.django.models import User
-from ludamus.pacts import UserType
+from ludamus.pacts.crowd import UserType
 from tests.integration.utils import assert_response
