+++ tests/integration/web/chronology/test_session_enroll_page.py
@@ -19,7 +19,7 @@ from ludamus.adapters.db.django.models import (
     UserEnrollmentConfig,
 )
 from ludamus.adapters.web.django.entities import SessionUserParticipationData
-from ludamus.pacts import UserDTO
+from ludamus.pacts.crowd import UserDTO
 from ludamus.pacts.legacy import NotificationKind
 from tests.integration.conftest import (
     AgendaItemFactory,
