+++ tests/integration/web/panel/test_proposals_page.py
@@ -18,8 +18,8 @@ from ludamus.pacts import (
     SessionListItemDTO,
     SessionStatus,
     TrackDTO,
-    UserDTO,
 )
+from ludamus.pacts.crowd import UserDTO
 from tests.integration.conftest import UserFactory
 from tests.integration.utils import assert_response
