+++ tests/integration/web/panel/test_track_create_page.py
@@ -7,7 +7,8 @@ from django.contrib import messages
 from django.urls import reverse
 
 from ludamus.adapters.db.django.models import Space, Track
-from ludamus.pacts import EventDTO, UserDTO
+from ludamus.pacts import EventDTO
+from ludamus.pacts.crowd import UserDTO
 from tests.integration.conftest import SpaceFactory, UserFactory
 from tests.integration.utils import assert_response
